Support stateful_path in control files
Adds a new option to os-apply-config control files called stateful_path.
The option is a boolean value defaulted to False. If True, the rendered
template's file path will be prefixed with the path from the
$STATEFUL_PATH environment variable.

This allows for some templates to be written under a configurable
stateful path directory (vs. a hardcoded /mnt/state), and some templates
to just be written under /.

Change-Id: I4c2ffbcc3183bf566209f6fc1446028847e6a59d